pyproject.toml
[tool.poetry]
name = "injective-py"
version = "1.11.0"
description = "Injective Python SDK, with Exchange API Client"
authors = ["Injective Labs <contact@injectivelabs.org>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://injectivelabs.org/"
repository = "https://github.com/InjectiveLabs/sdk-python"
documentation = "https://api.injective.exchange/"
keywords = ["injective", "blockchain", "cosmos", "injectivelabs"]
classifiers = [
    "Intended Audience :: Developers",
]
packages = [
    { include = "pyinjective" },
]
include = [
    { path = "pyinjective/proto/**/*.py" },
    { path = "*.ini" },
]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.4" # Version dependency due to https://github.com/InjectiveLabs/sdk-python/security/dependabot/18
bech32 = "*"
bip32 = "*"
ecdsa = "*"
eip712 = "*"
grpcio = "*"
grpcio-tools = "*"
hdwallets = "*"
mnemonic = "*"
protobuf = "^5.26.1"
requests = "*"
safe-pysha3 = "*"
websockets = "*"
web3 = "^7.0.0"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-grpc = "*"
requests-mock = "*"
pytest-cov = ">=4.1.0"
pytest-aioresponses = ">=0.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.4.0"
flakeheaven = ">=3.3.0"
isort = ">=5.12.0"
black = ">=23.9.1"
python-dotenv = ">=1.0.1"
importlib-metadata = "<5.0"


[tool.flakeheaven]
exclude = ["pyinjective/proto/*", ".idea/*", "*.md"]
max_line_length = 120


# list of plugins and rules for them
[tool.flakeheaven.plugins]
pycodestyle = ["+*", "-W503", "-E731"]
pyflakes = ["+*"]

[tool.flakeheaven.exceptions."tests/"]
pyflakes = ["-F811"]           # disable a plugin


[tool.isort]
line_length = 120
multi_line_output = 3
combine_as_imports = true
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["pyinjective/proto/*", ".idea/*"]


[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^pyinjective/proto/.*
  | \.idea/.*
)
'''

[tool.coverage.run]
branch = true
source_pkgs = ["pyinjective"]
omit = ["pyinjective/proto/*"]

[tool.coverage.report]
skip_empty = true
fail_under = 65
precision = 2

[tool.pytest.ini_options]
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
